requirements.txt
Django==3.2.8
djangorestframework==3.12.4
requests==2.26.0
# psycopg2-binary==2.8.5
gunicorn==20.0.4
